Don't report ECONNABORT or ECONNRESET as read errors
If we're just idle waiting for bytes to become available to read and the
connection is forcibly reset or aborted, just treat it as if it were closed
normally. This lets us bubble back the actual number of bytes transmitted.

We are still treating these as errors if we encounter them while actively trying
to write to the host/client.

Should resolve https://github.com/mqudsi/tcpproxy/issues/3